style(turbopack): clippy --fix (vercel/turbo#3014)
Easy clippy improvements via `clippy --fix`.

+++ crates/turbo-tasks-memory/src/viz/mod.rs
@@ -82,7 +82,7 @@ fn get_max_values_internal(depth: u32, node: &GroupTree) -> MaxValues {
         max_total_current_duration = max(max_total_current_duration, s.total_current_duration);
         max_total_update_duration = max(max_total_update_duration, s.total_update_duration);
         if let Some((total_duration, executions)) = s.total_duration.zip(s.executions) {
-            let avg_duration = total_duration / executions as u32;
+            let avg_duration = total_duration / executions;
             max_avg_duration = max_avg_duration
                 .map(|max_avg_duration| max(max_avg_duration, avg_duration))
                 .or(Some(avg_duration));
